refactor: move global response schema types to dedicated entrypoint
