+++ projects/api/src/contracts/_internal/response/likeResponseSchema.ts
@@ -1,5 +1,5 @@
 import { z } from '../z.ts';
-import { profileResponseSchema } from './userProfileResponseSchema.ts';
+import { profileResponseSchema } from './profileResponseSchema.ts';
 
 export const likeResponseSchema = z.object({
   liked_at: z.string(),
