--- projects/api/src/contracts/_internal/response/movieAnticipatedResponseSchema.ts
@@ -1,7 +1,0 @@
-import { z } from '../z.ts';
-import { movieResponseSchema } from './movieResponseSchema.ts';
-
-export const movieAnticipatedResponseSchema = z.object({
-  list_count: z.number(),
-  movie: movieResponseSchema,
-});
