--- projects/api/src/contracts/_internal/response/movieTrendingResponseSchema.ts
@@ -1,7 +1,0 @@
-import { z } from '../z.ts';
-import { movieResponseSchema } from './movieResponseSchema.ts';
-
-export const movieTrendingResponseSchema = z.object({
-  watchers: z.number(),
-  movie: movieResponseSchema,
-});
