--- projects/api/src/contracts/_internal/response/showAnticipatedResponseSchema.ts
@@ -1,7 +1,0 @@
-import { z } from '../z.ts';
-import { showResponseSchema } from './showResponseSchema.ts';
-
-export const showAnticipatedResponseSchema = z.object({
-  list_count: z.number(),
-  show: showResponseSchema,
-});
