--- projects/api/src/contracts/_internal/response/showTrendingResponseSchema.ts
@@ -1,7 +1,0 @@
-import { z } from '../z.ts';
-import { showResponseSchema } from './showResponseSchema.ts';
-
-export const showTrendingResponseSchema = z.object({
-  watchers: z.number(),
-  show: showResponseSchema,
-});
