+++ projects/api/src/contracts/movies/_internal/response/movieAnticipatedResponseSchema.ts
@@ -1,0 +1,7 @@
+import { movieResponseSchema } from '../../../_internal/response/movieResponseSchema.ts';
+import { z } from '../../../_internal/z.ts';
+
+export const movieAnticipatedResponseSchema = z.object({
+  list_count: z.number(),
+  movie: movieResponseSchema,
+});
