+++ projects/api/src/contracts/movies/_internal/response/movieTrendingResponseSchema.ts
@@ -1,0 +1,7 @@
+import { movieResponseSchema } from '../../../_internal/response/movieResponseSchema.ts';
+import { z } from '../../../_internal/z.ts';
+
+export const movieTrendingResponseSchema = z.object({
+  watchers: z.number(),
+  movie: movieResponseSchema,
+});
