+++ projects/api/src/contracts/recommendations/_internal/response/userProfileWithNotesSchema.ts
@@ -1,4 +1,4 @@
-import { profileResponseSchema } from '../../../_internal/response/userProfileResponseSchema.ts';
+import { profileResponseSchema } from '../../../_internal/response/profileResponseSchema.ts';
 import { z } from '../../../_internal/z.ts';
 
 export const userProfileWithNotesSchema = profileResponseSchema
