+++ projects/api/src/contracts/shows/_internal/response/showAnticipatedResponseSchema.ts
@@ -1,0 +1,7 @@
+import { showResponseSchema } from '../../../_internal/response/showResponseSchema.ts';
+import { z } from '../../../_internal/z.ts';
+
+export const showAnticipatedResponseSchema = z.object({
+  list_count: z.number(),
+  show: showResponseSchema,
+});
