+++ projects/api/src/contracts/sync/_internal/response/ratingsResponseSchema.ts
@@ -1,7 +1,7 @@
 import { z } from '../../../_internal/z.ts';
 import { ratingsParamSchema } from '../request/ratingsParamSchema.ts';
 
-export const ratingsResponseSchema = z.object({
+export const ratingsSyncResponseSchema = z.object({
   added: z.object({
     movies: z.number(),
     shows: z.number(),
