+++ projects/api/src/contracts/users/_internal/response/friendResponseSchema.ts
@@ -1,4 +1,4 @@
-import { profileResponseSchema } from '../../../_internal/response/userProfileResponseSchema.ts';
+import { profileResponseSchema } from '../../../_internal/response/profileResponseSchema.ts';
 import { z } from '../../../_internal/z.ts';
 
 export const friendResponseSchema = z.object({
